dashboard/backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
pyyaml==6.0.1
httpx==0.25.2
asyncssh==2.14.0
redis==5.0.1

